requirements.txt
tensorflow==2.18.0
numpy
pillow
streamlit
